+++ script/pipeline_1_train_xgboost.cpp
@@ -14,12 +14,12 @@ using namespace std;
 
 const int min_leaf = 1;
 const int max_bin = 32;
-const double learning_rate = 0.3;
-const double lam = 0.0;
-const double const_gamma = 0.0;
-const double eps = 1.0;
-const double min_child_weight = -1 * numeric_limits<double>::infinity();
-const double subsample_cols = 0.8;
+const float learning_rate = 0.3;
+const float lam = 0.0;
+const float const_gamma = 0.0;
+const float eps = 1.0;
+const float min_child_weight = -1 * numeric_limits<float>::infinity();
+const float subsample_cols = 0.8;
 
 string folderpath;
 string fileprefix;
@@ -82,8 +82,8 @@ int main(int argc, char *argv[])
     int num_row_train, num_row_val, num_col, num_party;
     int num_nan_cell = 0;
     scanf("%d %d %d", &num_row_train, &num_col, &num_party);
-    vector<vector<double>> X_train(num_row_train, vector<double>(num_col));
-    vector<double> y_train(num_row_train);
+    vector<vector<float>> X_train(num_row_train, vector<float>(num_col));
+    vector<float> y_train(num_row_train);
     vector<XGBoostParty> parties(num_party);
 
     int temp_count_feature = 0;
@@ -92,13 +92,13 @@ int main(int argc, char *argv[])
         int num_col = 0;
         scanf("%d", &num_col);
         vector<int> feature_idxs(num_col);
-        vector<vector<double>> x(num_row_train, vector<double>(num_col));
+        vector<vector<float>> x(num_row_train, vector<float>(num_col));
         for (int j = 0; j < num_col; j++)
         {
             feature_idxs[j] = temp_count_feature;
             for (int k = 0; k < num_row_train; k++)
             {
-                scanf("%lf", &x[k][j]);
+                scanf("%f", &x[k][j]);
                 if (use_missing_value && x[k][j] == -1)
                 {
                     x[k][j] = nan("");
@@ -112,16 +112,16 @@ int main(int argc, char *argv[])
         parties[i] = party;
     }
     for (int j = 0; j < num_row_train; j++)
-        scanf("%lf", &y_train[j]);
+        scanf("%f", &y_train[j]);
 
     scanf("%d", &num_row_val);
-    vector<vector<double>> X_val(num_row_val, vector<double>(num_col));
-    vector<double> y_val(num_row_val);
+    vector<vector<float>> X_val(num_row_val, vector<float>(num_col));
+    vector<float> y_val(num_row_val);
     for (int i = 0; i < num_col; i++)
     {
         for (int j = 0; j < num_row_val; j++)
         {
-            scanf("%lf", &X_val[j][i]);
+            scanf("%f", &X_val[j][i]);
             if (use_missing_value && X_val[j][i] == -1)
             {
                 X_val[j][i] = nan("");
@@ -129,7 +129,7 @@ int main(int argc, char *argv[])
         }
     }
     for (int j = 0; j < num_row_val; j++)
-        scanf("%lf", &y_val[j]);
+        scanf("%f", &y_val[j]);
 
     std::ofstream result_file;
     string result_filepath = folderpath + "/" + fileprefix + "_result.ans";
@@ -155,7 +155,7 @@ int main(int argc, char *argv[])
     start = chrono::system_clock::now();
     clf.fit(parties, y_train);
     end = chrono::system_clock::now();
-    double elapsed = chrono::duration_cast<chrono::milliseconds>(end - start).count();
+    float elapsed = chrono::duration_cast<chrono::milliseconds>(end - start).count();
     printf("Training is complete %f [ms] seed=%s\n", elapsed, fileprefix.c_str());
 
     for (int i = 0; i < clf.logging_loss.size(); i++)
@@ -169,10 +169,10 @@ int main(int argc, char *argv[])
         result_file << clf.estimators[i].print(true, true).c_str() << "\n";
     }
 
-    vector<double> predict_proba_train = clf.predict_proba(X_train);
+    vector<float> predict_proba_train = clf.predict_proba(X_train);
     vector<int> y_true_train(y_train.begin(), y_train.end());
     result_file << "Train AUC," << roc_auc_score(predict_proba_train, y_true_train) << "\n";
-    vector<double> predict_proba_val = clf.predict_proba(X_val);
+    vector<float> predict_proba_val = clf.predict_proba(X_val);
     vector<int> y_true_val(y_val.begin(), y_val.end());
     result_file << "Val AUC," << roc_auc_score(predict_proba_val, y_true_val) << "\n";
     result_file.close();
